Check if Jetpack.css is really needed.
This runs through all the Jetpack registered styles and checks how many are enqueued before enqueuing the concatenated file.

If two or more are enqueued then the concatenated file is loaded. In the days of H2 this could be smarter but eh?

If zero or one file are enqueued then the concatenated file is not loaded because the original approach will result in a smaller transfer.